% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_geocoding_sequence_files.R
\name{export_geocoding_sequence_files}
\alias{export_geocoding_sequence_files}
\title{Create files for batchgeocoding.}
\usage{
export_geocoding_sequence_files(addresses,
  file_path = "c:/temp/geocode/", max_rows_per_file = 50000)
}
\arguments{
\item{addresses}{a \code{data.frame} containing one column named address and additional columns as desired.}

\item{file_path}{a \code{character} destination for sequence files and the run.bat used to execute docker based geocoding}

\item{max_rows_per_file}{a \code{integer} specifying the maximum number of rows per file to be geocoded.}
}
\description{
Convert a large data frame containing and address column
into multiple files for batch geocoding
}
