% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoframework.R
\docType{package}
\name{geoframework-package}
\alias{geoframework}
\alias{geoframework-package}
\title{geoframework: Geocoding, Routing, Drive Distance, Isochrones, Census Boundaries, and Socioeconomic Data}
\description{
A collection of functions, data sets, and references to aid in
    performing geospatial analysis and visualization in R.
}
\details{
This package is intended to leverage open source geospatial solutions that
operate behind a firewall. There is no guarantee that all of the referenced
code is open source. There is no guarantee that no data is being sent to an
external network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/geoframework}
  \item Report bugs at \url{https://github.com/johnaclouse/geoframework/issues}
}

}
\keyword{internal}
